Split the micro_booter + rename into kernel_tests
