--- src/components/implementation/no_interface/vkernel/Makefile
@@ -1,9 +1,0 @@
-ASM_OBJS=inv.o invvk.o
-COMPONENT=vkernel_boot.o
-INTERFACES=
-DEPENDENCIES=
-IF_LIB=
-ADDITIONAL_LIBS=-lcobj_format $(LIBSLRAW) -lsl_mod_fprr -lsl_thd_static_backend
-
-include ../../Makefile.subsubdir
-MANDITORY_LIB=simple_stklib.o
