--- src/components/implementation/no_interface/vkernel/inv.S
@@ -1,1 +1,0 @@
-../../tests/micro_booter/inv.S
