+++ src/components/implementation/tests/kernel_tests/Makefile
@@ -1,0 +1,10 @@
+C_OBJS=kernel_test_booter.o k_test_async.o k_test_captbl.o k_test_mem.o k_test_tcap.o k_test_thd.o k_test_inv.o k_perf_tests.o
+ASM_OBJS=cos_asm_scheduler.o inv.o
+COMPONENT=kernel_tests.o
+INTERFACES=
+DEPENDENCIES=
+IF_LIB=
+ADDITIONAL_LIBS=-lcobj_format -lcos_kernel_api -lcos_unit
+
+include ../../Makefile.subsubdir
+MANDITORY_LIB=simple_stklib.o
