--- src/components/implementation/tests/micro_booter/Makefile
@@ -1,10 +1,0 @@
-C_OBJS=micro_booter.o mb_tests.o mb_perf_tests.o
-ASM_OBJS=cos_asm_scheduler.o inv.o
-COMPONENT=micro_boot.o
-INTERFACES=
-DEPENDENCIES=
-IF_LIB=
-ADDITIONAL_LIBS=-lcobj_format -lcos_kernel_api -lcos_expect
-
-include ../../Makefile.subsubdir
-MANDITORY_LIB=simple_stklib.o
