+++ src/components/implementation/tests/micro_xcores/Makefile
@@ -4,7 +4,7 @@ COMPONENT=micro_xcore.o
 INTERFACES=
 DEPENDENCIES=
 IF_LIB=
-ADDITIONAL_LIBS=-lcobj_format -lcos_kernel_api -lcos_expect
+ADDITIONAL_LIBS=-lcobj_format -lcos_kernel_api -lcos_unit
 
 include ../../Makefile.subsubdir
 MANDITORY_LIB=simple_stklib.o
