+++ src/platform/i386/runscripts/kernel_tests.sh
@@ -1,0 +1,4 @@
+#!/bin/sh
+
+cp kernel_tests.o llboot.o
+./cos_linker "llboot.o, :" ./gen_client_stub
